create organised directory structure for project
