+++ public/customFigureBuilder.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset='UTF-8'>
   <title>Title</title>
-  <script defer src='customFigureBuilder.js'></script>
+  <script defer src='../src/js/pages/customFigureBuilder.js'></script>
 </head>
 <body>
