+++ public/custom_figure.html
@@ -5,10 +5,13 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Neue Kennzahl erstellen</title>
-  <link rel="stylesheet" href="custome_figure.css" />
-  <link rel="stylesheet" href="sidebar.css" />
-  <link rel="stylesheet" href="infobox.css">
-  <script type="module" src="customFigureBuilder.js" defer></script>
+  <link rel="stylesheet" href="../src/styles/custome_figure.css" />
+  <link rel="stylesheet" href="../src/styles/sidebar.css" />
+  <link rel="stylesheet" href="../src/styles/infobox.css">
+  <script type="module" src="../src/js/pages/customFigureBuilder.js" defer></script>
+  <script type="module" src="../src/js/components/customKeyFigureSidebar.js" defer></script>
+  <script type="module" src="../src/script.js"></script>
+
 </head>
 <body>
 <div class="content-wrapper">
@@ -62,7 +65,6 @@
 </div>
 
 </div>
-<script src="script.js"></script>
 </body>
 
 </body>
