+++ public/import.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Import</title>
   <script src="https://cdn.tailwindcss.com"></script>
-  <link rel="stylesheet" href="styles.css">
+  <link rel="stylesheet" href="../src/styles/styles.css">
 </head>
 <body class="bg-gray-100 flex flex-col items-center justify-center h-screen">
 
@@ -21,6 +21,6 @@
   <button id="upload" class="upload-button">Hochladen</button>
 </div>
 
-<script src="script.js"></script>
+<script src="../src/script.js"></script>
 </body>
 </html>
