+++ public/index.html
@@ -5,12 +5,12 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Kennzahlen Cockpit</title>
   <script src="https://cdn.tailwindcss.com"></script>
-  <link rel="stylesheet" href="styles.css" />
-  <link rel="stylesheet" href="sidebar.css">
-  <link rel="stylesheet" href="infobox.css">
-  <script type="module" src="companySidebar.js" defer></script>
-  <script src="loadCompanyData.js" defer></script>
-  <script type="module" src="script.js" defer></script>
+  <link rel="stylesheet" href="../src/styles/styles.css" />
+  <link rel="stylesheet" href="../src/styles/sidebar.css">
+  <link rel="stylesheet" href="../src/styles/infobox.css">
+  <script type="module" src="../src/js/components/companySidebar.js" defer></script>
+  <script src="../src/js/keyFigureData/loadCompanyData.js" defer></script>
+  <script type="module" src="../src/script.js" defer></script>
 </head>
