+++ public/login.html
@@ -5,9 +5,9 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Login</title>
   <script src="https://cdn.tailwindcss.com"></script>
-  <script type="module" src="script.js"></script>
-  <link rel="stylesheet" href="styles.css">
-  <link rel="stylesheet" href="infobox.css">
+  <script type="module" src="../src/script.js"></script>
+  <link rel="stylesheet" href="../src/styles/styles.css">
+  <link rel="stylesheet" href="../src/styles/infobox.css">
 </head>
 <body class="bg-gray-100 flex justify-center items-center h-screen">
