+++ src/js/keyFigureData/loadCompanyData.js
