+++ src/js/utils/serverResponseHandling.js
