+++ src/styles/infobox.css
